mlfow

+++ backend/ml/models/components/normalization.py
@@ -3,10 +3,10 @@ import torch.nn as nn
 
 
 class LayerNorm(nn.Module):
-    def __init__(self, d_model, eps=1e-6):
+    def __init__(self, d_model, eps=1e-6, dtype=torch.float32):
         super().__init__()
-        self.scale = nn.Parameter(torch.ones(d_model))
-        self.shift = nn.Parameter(torch.zeros(d_model))
+        self.scale = nn.Parameter(torch.ones(d_model, dtype=dtype))
+        self.shift = nn.Parameter(torch.zeros(d_model, dtype=dtype))
         self.eps = eps
 
     def forward(self, x):
@@ -17,9 +17,9 @@ class LayerNorm(nn.Module):
 
 
 class RMSNorm(nn.Module):
-    def __init__(self, d_model, eps=1e-6):
+    def __init__(self, d_model, eps=1e-6, dtype=torch.float32):
         super().__init__()
-        self.scale = nn.Parameter(torch.ones(d_model))
+        self.scale = nn.Parameter(torch.ones(d_model, dtype=dtype))
         self.eps = eps
 
     def forward(self, x):
